+++ src/service/auth/request/index.ts
@@ -1,0 +1,14 @@
+export interface SignInData {
+  provider: string;
+  email: string;
+  oAuthToken: string;
+}
+
+export interface SignUpData {
+  provider: string;
+  oAuthToken: string;
+  email: string;
+  nickname: string;
+  thirdPartyProfileUrl?: string;
+  belonging?: string;
+}
