+++ src/service/belonging/response/index.ts
@@ -1,0 +1,8 @@
+import { Belonging } from 'types/common';
+
+export interface BelongingData {
+  belongingId: number;
+  belongingType: Belonging;
+  name: string;
+  imageUrl: string;
+}
