+++ src/service/interceptor/index.ts
@@ -1,18 +1,6 @@
-import { ApiResponse, RequestInitWithAuth } from '../service';
-import { handleResponseByCode } from './response';
+import { RequestConfigWithResponse } from '@/service/types';
 
-export interface RequestConfigWithResponse<T> extends RequestInitWithAuth {
-  url: string;
-  method: string;
-  request: <T = unknown>(
-    method: string,
-    url: string,
-    data?: unknown,
-    config?: RequestInitWithAuth
-  ) => Promise<T>;
-  baseURL: string;
-  response?: ApiResponse<T>;
-}
+import { handleResponseByCode } from './response';
 
 export type InterceptorFunction<T = unknown> = (
   config: RequestConfigWithResponse<T>
