+++ src/service/types.ts
@@ -1,0 +1,22 @@
+export interface ApiResponse<T> {
+  code: number;
+  message: string;
+  value: T;
+}
+
+export interface RequestInitWithAuth extends RequestInit {
+  includeAuth?: boolean;
+}
+
+export interface RequestConfigWithResponse<T> extends RequestInitWithAuth {
+  url: string;
+  method: string;
+  request: <T = unknown>(
+    method: string,
+    url: string,
+    data?: unknown,
+    config?: RequestInitWithAuth
+  ) => Promise<T>;
+  baseURL: string;
+  response?: ApiResponse<T>;
+}
